pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-application-autopilot"
version = "0.1.0"
description = "AI-powered job application agent — ATS checker, resume tailor, multi-platform search via MCP"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Mrityunjay"},
]

# Core — required for all modes. No NLP library needed.
dependencies = [
    "fastmcp>=2.0.0",
    "typer[all]>=0.9.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "docling>=2.0.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
# Standalone CLI with Claude / Anthropic API
anthropic = ["anthropic>=0.40.0"]

# Standalone CLI with Google Gemini API
google = ["google-generativeai>=0.8.0"]

# Standalone CLI with local models (free, no key, privacy-first)
ollama = ["ollama>=0.4.0"]

# Job platform login (not needed for Antigravity users)
browser = ["playwright>=1.40.0"]

# Install everything
all = ["anthropic>=0.40.0", "google-generativeai>=0.8.0", "ollama>=0.4.0", "playwright>=1.40.0"]

[project.scripts]
job-autopilot = "cli:app"
job-autopilot-server = "server:mcp.run"

[tool.setuptools.packages.find]
include = ["src*"]
